refactor(floating-edge-example): cleanup
